# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Sistema POS",
        "description": "Documentación de la API del Sistema POS",
        "contact": {
            "email": "admin@sistema.com"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/proveedores": {
            "get": {
                "tags": [
                    "Proveedores"
                ],
                "summary": "Listar todos los proveedores",
                "operationId": "7f6ed20dc86f1682aa7e31d71afa37df",
                "responses": {
                    "200": {
                        "description": "Lista de proveedores obtenida exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer"
                                                    },
                                                    "nombre": {
                                                        "type": "string"
                                                    },
                                                    "tipo_documento": {
                                                        "type": "string",
                                                        "enum": [
                                                            "RNC",
                                                            "CEDULA"
                                                        ]
                                                    },
                                                    "documento": {
                                                        "type": "string"
                                                    },
                                                    "email": {
                                                        "type": "string",
                                                        "nullable": true
                                                    },
                                                    "telefono": {
                                                        "type": "string",
                                                        "nullable": true
                                                    },
                                                    "direccion": {
                                                        "type": "string"
                                                    },
                                                    "condicion_pago_id": {
                                                        "type": "integer"
                                                    },
                                                    "is_activo": {
                                                        "type": "boolean"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Proveedores"
                ],
                "summary": "Crear un nuevo proveedor",
                "operationId": "a1046f3d7a3c804651bf420761995e57",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "nombre",
                                    "tipo_documento",
                                    "documento",
                                    "direccion",
                                    "condicion_pago_id"
                                ],
                                "properties": {
                                    "nombre": {
                                        "type": "string",
                                        "maxLength": 100
                                    },
                                    "tipo_documento": {
                                        "type": "string",
                                        "enum": [
                                            "RNC",
                                            "CEDULA"
                                        ]
                                    },
                                    "documento": {
                                        "type": "string",
                                        "maxLength": 20
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "maxLength": 100,
                                        "nullable": true
                                    },
                                    "telefono": {
                                        "type": "string",
                                        "maxLength": 20,
                                        "nullable": true
                                    },
                                    "direccion": {
                                        "type": "string",
                                        "maxLength": 255
                                    },
                                    "condicion_pago_id": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Proveedor creado exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "message": {
                                            "type": "string"
                                        },
                                        "data": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer"
                                                },
                                                "nombre": {
                                                    "type": "string"
                                                },
                                                "tipo_documento": {
                                                    "type": "string"
                                                },
                                                "documento": {
                                                    "type": "string"
                                                },
                                                "email": {
                                                    "type": "string",
                                                    "nullable": true
                                                },
                                                "telefono": {
                                                    "type": "string",
                                                    "nullable": true
                                                },
                                                "direccion": {
                                                    "type": "string"
                                                },
                                                "condicion_pago_id": {
                                                    "type": "integer"
                                                },
                                                "is_activo": {
                                                    "type": "boolean"
                                                },
                                                "condicionPago": {
                                                    "properties": {
                                                        "id": {
                                                            "type": "integer"
                                                        },
                                                        "nombre": {
                                                            "type": "string"
                                                        },
                                                        "dias": {
                                                            "type": "integer"
                                                        }
                                                    },
                                                    "type": "object"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Error de validación"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
        "/api/proveedores/{proveedor}/balance": {
            "get": {
                "tags": [
                    "Proveedores"
                ],
                "summary": "Obtener el balance de un proveedor",
                "operationId": "0ce306d23fce378c2e08ae378fcc12c1",
                "parameters": [
                    {
                        "name": "proveedor",
                        "in": "path",
                        "description": "ID del proveedor",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Balance del proveedor obtenido exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "data": {
                                            "properties": {
                                                "total_compras": {
                                                    "type": "number",
                                                    "format": "float"
                                                },
                                                "total_pagado": {
                                                    "type": "number",
                                                    "format": "float"
                                                },
                                                "balance_pendiente": {
                                                    "type": "number",
                                                    "format": "float"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Proveedor no encontrado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "sanctum",
                "scheme": "bearer"
            }
        }
    },
...
}